Refactor cache-related sources
