+++ server-infrastructure/src/main/java/io/joyoungc/infrastructure/cache/model/MemberCache.java
@@ -1,14 +1,16 @@
-package io.joyoungc.api.order.request;
+package io.joyoungc.infrastructure.cache.model;
 
 import io.joyoungc.domain.common.constant.Constants;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.io.Serial;
 import java.io.Serializable;
 
 @Getter
 @Setter
-public class OrderCacheResponse implements Serializable {
+public class MemberCache implements Serializable {
+    @Serial
     private static final long serialVersionUID = Constants.SERIAL_VERSION_UID;
     private Long id;
     private String name;
