+++ server-infrastructure/src/main/java/io/joyoungc/infrastructure/persistence/mapper/MemberMapper.java
@@ -2,6 +2,7 @@ package io.joyoungc.infrastructure.persistence.mapper;
 
 import io.joyoungc.domain.member.Member;
 import io.joyoungc.domain.order.Order;
+import io.joyoungc.infrastructure.cache.model.MemberCache;
 import io.joyoungc.infrastructure.persistence.entity.MemberEntity;
 import org.mapstruct.AfterMapping;
 import org.mapstruct.Mapper;
@@ -19,6 +20,8 @@ public interface MemberMapper {
     Member toMember(MemberEntity memberEntity);
 
     List<Member> toMemberList(List<MemberEntity> memberEntities);
+
+    MemberCache toMemberCache(MemberEntity memberEntity);
 
 
     @AfterMapping
